Projects/User_Registeration_management/requirements.txt
fastapi
uvicorn
pydantic
python-jose
bcrypt
motor
sqlalchemy
pymysql
email-validator
python-dotenv
pyjwt
